# docs/topics/gamification_openapi_2.self-contained.yaml
...
openapi: "3.0.3"
info:
  title: "gamification API"
  description: "gamification API"
  version: "1.0.0"
servers:
  - url: "https://api.tc-syd-02.insa-lyon.fr"
    description: "Production server"
paths:
...
  /api/grade/all/{sessionId}:
    get:
      summary: "Get all grades for a session"
      description: "Get all grades for a session. Returns a list of grades. Only available to teachers/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "sessionId"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    progressionId:
                      type: string
                      example: "65a7abe1ef63234c1129d7f9"
                    studentName:
                      type: string
                      example: "Student A"
                    grade:
                      type: number
                      example: 10
                    bonus:
                      type: number
                      example: 0
                    level:
                      type: integer
                      example: 1
                    gradeOverride:
                      type: boolean
                      example: false
                    gradeComment:
                      type: string
                      example: "Comment"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/grade/resetGrade/{progressionId}:
    post:
      summary: "Remove grade override for student"
      description: "Remove grade override for student. Only available to teachers/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "progressionId"
          in: "path"
          schema:
            type: string
          description: The progression id
          required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Grade reset"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Progression not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/grade/setGrade/{progressionId}:
    post:
      summary: "Set grade override for student"
      description: "Set grade override for student. Only available to teachers/admin - Requires authentication"
      parameters:
        - name: "progressionId"
          in: "path"
          schema:
            type: string
          description: The progression id
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                grade:
                  type: number
                  example: 10
                comment:
                  type: string
                  example: "Comment"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Grade overriden"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Grade is missing"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Progression not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/grade/setLevelGrade/{sessionId}:
    post:
      summary: "Set level index grade for session"
      description: "Set level index grade for session. Returns the list of non-modified grades. Only available to teachers/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "sessionId"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: integer
                  example: 1
                grade:
                  type: number
                  example: 10
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    progressionId:
                      type: string
                      example: "65a7abe1ef63234c1129d7f9"
                    studentName:
                      type: string
                      example: "Student A"
                    grade:
                      type: number
                      example: 10
                    level:
                      type: integer
                      example: 1
                    gradeOverriden:
                      type: boolean
                      example: false
                    gradeComment:
                      type: string
                      example: "Comment"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Level or grade is missing"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/grade/removeLevelGrade/{sessionId}:
    post:
      summary: "Reset level index grade for session"
      description: "Reset level index grade for session. Returns the list of non-modified grades. Only available to teachers/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "sessionId"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    progressionId:
                      type: string
                      example: "65a7abe1ef63234c1129d7f9"
                    studentName:
                      type: string
                      example: "Student A"
                    grade:
                      type: number
                      example: 10
                    level:
                      type: integer
                      example: 1
                    gradeOverriden:
                      type: boolean
                      example: false
                    gradeComment:
                      type: string
                      example: "Comment"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/grade/editStudentLevel/{progressionId}:
    post:
      summary: "Edit student level"
      description: "Edit the student's level. Only available to teachers/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "progressionId"
          in: "path"
          schema:
            type: string
          description: The progression id
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: integer
                  example: 1
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Level edited"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Level is missing"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Progression not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/grade/students/{sessionid}:
    get:
      summary: "Get students for autocomplete"
      description: "Get students for autocomplete from searchString. Only available to teachers/admin - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "sessionid"
          in: "path"
          schema:
            type: string
          description: The session id
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                searchString:
                  type: string
                  example: "searchString"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherUser"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...
components:
  schemas:
...
    TeacherUser:
      description: "A teacher user"
      type: "object"
      properties:
        id:
          description: "The user's id"
          type: "string"
        name:
          description: "The user's name"
          type: "string"
        surname:
          description: "The user's surname"
          type: "string"
        type:
          description: "The user's type"
          type: "string"
          enum:
            - "student"
            - "teacher"
            - "admin"
        tps:
          description: "The user's available tps"
          type: "array"
          items:
            type: "string"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid